implement state pushdown logic for optimized query handling and add related utility functions
