+++ packages/engine/src/sql/steps/mod.rs
@@ -12,6 +12,7 @@ pub mod lix_state_view_read;
 pub mod lix_state_view_write;
 pub mod lix_version_view_read;
 pub mod lix_version_view_write;
+pub(crate) mod state_pushdown;
 pub mod stored_schema;
 pub mod vtable_read;
 pub mod vtable_write;
